♻ refactor : 프로젝트 파일 구조 변경 #119 하나의 디렉토리에 들어있던 것들을 각 역할에 맞게 분리
