+++ frontend/public/assets/achievements/.filename
@@ -1,0 +1,4 @@
+<achievementId>.png와 같이 명명
+
+디렉토리 생성하기 위한 파일입니다.
+제거될 예정
