+++ frontend/src/App.tsx
@@ -1,15 +1,14 @@
-import React from 'react';
 import { BrowserRouter, NavLink, Routes, Route } from 'react-router-dom';
-import Ranks from './components/Ranks';
-import Profile from './components/Profile';
-import './App.css';
+import Ranks from './views/Ranks';
+import Profile from './views/Profile';
+import './style/App.css';
 
 function App() {
   return (
     <BrowserRouter>
       <header>headerheaderheaderheaderheader</header>
       <nav>
-        {/* 위에 로고 */}
+        {/* TODO :위에 로고 */}
         <ul>
           <li>
             <NavLink to="/profile">Profile</NavLink>
@@ -24,7 +23,7 @@ function App() {
             <NavLink to="/game">Game</NavLink>
           </li>
         </ul>
-        {/* 아래 로그아웃 */}
+        {/* TODO : 아래 로그아웃 */}
       </nav>
       <main>
         <Routes>
