--- frontend/src/components/Ranks.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Ranks() {
-  return (
-    <div>
-      <h1>RANKS</h1>
-    </div>
-  );
-}
-
-export default Ranks;
